Moved classes for tile table data into own directory
