+++ src/migration/BatchTableClassProperties.ts
@@ -2,7 +2,7 @@ import { ClassProperty } from "../structure/Metadata/ClassProperty";
 
 import { TileFormatError } from "../tileFormats/TileFormatError";
 
-import { TileTableData } from "./TileTableData";
+import { TileTableData } from "../tileTableData/TileTableData";
 import { TypeDetection } from "./TypeDetection";
 
 /**
