+++ src/tileTableData/AttributeCompression.ts
