+++ src/tileTableData/VecMath.ts
