+++ src/tilesetProcessing/TilesetJsonCreator.ts
@@ -24,9 +24,9 @@ import { BoundingVolumes } from "./BoundingVolumes";
 import { BoundingBox3D } from "./BoundingVolumes";
 import { Point3D } from "./BoundingVolumes";
 
-import { TileTableData } from "../migration/TileTableData";
+import { TileTableData } from "../tileTableData/TileTableData";
 import { TileTableDataI3dm } from "../migration/TileTableDataI3dm";
-import { VecMath } from "../migration/VecMath";
+import { VecMath } from "../tileTableData/VecMath";
 
 const DEFAULT_LEAF_GEOMETRIC_ERROR = 512;
 const DEFAULT_TILESET_GEOMETRIC_ERROR = 4096;
